+++ sshd-core/src/main/java/org/apache/sshd/client/session/AbstractClientSession.java
@@ -58,7 +58,6 @@ import org.apache.sshd.common.channel.PtyChannelConfigurationHolder;
 import org.apache.sshd.common.cipher.BuiltinCiphers;
 import org.apache.sshd.common.cipher.CipherNone;
 import org.apache.sshd.common.config.keys.KeyUtils;
-import org.apache.sshd.common.config.keys.OpenSshCertificate;
 import org.apache.sshd.common.forward.Forwarder;
 import org.apache.sshd.common.future.DefaultKeyExchangeFuture;
 import org.apache.sshd.common.future.KeyExchangeFuture;
@@ -618,27 +617,10 @@ public abstract class AbstractClientSession extends AbstractSession implements C
             remoteAddress = targetServerAddress.toInetSocketAddress();
         }
 
-        boolean verified = false;
-        if (serverKey instanceof OpenSshCertificate) {
-            // check if we trust the CA
-            verified = serverKeyVerifier.verifyServerKey(this, remoteAddress, ((OpenSshCertificate) serverKey).getCaPubKey());
-            if (log.isDebugEnabled()) {
-                log.debug("checkCA({}) key={}-{}, verified={}",
-                        this, KeyUtils.getKeyType(serverKey), KeyUtils.getFingerPrint(serverKey), verified);
-            }
-
-            if (!verified) {
-                // fallback to actual public host key
-                serverKey = ((OpenSshCertificate) serverKey).getCertPubKey();
-            }
-        }
-
-        if (!verified) {
-            verified = serverKeyVerifier.verifyServerKey(this, remoteAddress, serverKey);
-            if (log.isDebugEnabled()) {
-                log.debug("checkKeys({}) key={}-{}, verified={}",
-                        this, KeyUtils.getKeyType(serverKey), KeyUtils.getFingerPrint(serverKey), verified);
-            }
+        boolean verified = serverKeyVerifier.verifyServerKey(this, remoteAddress, serverKey);
+        if (log.isDebugEnabled()) {
+            log.debug("checkKeys({}) key={}-{}, verified={}", this, KeyUtils.getKeyType(serverKey),
+                    KeyUtils.getFingerPrint(serverKey), verified);
         }
 
         if (!verified) {